[steps] Extract parseBuildStepInputValueTypeName into a shared helper
